Final Cleanup

+++ Pengo/WallManager.h
@@ -1,5 +1,6 @@
 ﻿#pragma once
 #include "GameObject.h"
+#include "glm/vec2.hpp"
 
 class WallManager final : public GameObject
 {
@@ -19,7 +20,7 @@ public:
 	//Used for physics & networking
 	virtual void FixedUpdate(float fixedTimeMStep) override;
 
-	//Called after the Update()c
+	//Called after the Update()
 	//Used for camera and deletion of objects -> Deletion could be handled by the double buffer pattern
 	virtual void LateUpdate() override;
 
@@ -35,7 +36,7 @@ private:
 	glm::vec2 m_HorizontalWallSize{ 224,8 };
 	glm::vec2 m_VerticalWallSize{ 8,256 };
 
-	int m_WallOffset{ 50 };
+	float m_WallOffset{ 50 };
 
 	float m_TimeUntilMovingStateIsOver{ 3.f };
